+++ Core/Interfaces/Calculations/IResponse.cs
@@ -5,7 +5,7 @@ namespace StockTracker.Core.Interfaces.Calculations
     {
         public DateTime ActivityDate {get; set;}
 
-        public float GetFloatValue(string propertyName);
-        public void SetFloatValue(string propertyName, float value);
+        public decimal GetDecimalValue(string propertyName);
+        public void SetDecimalValue(string propertyName, decimal value);
     }
 }
